refactor(auth): update stage repository import and enhance user stage caching
